Updated maxCosignedAccountsPerAccount from uint8 to uint32

+++ plugins/txes/multisig/src/config/MultisigConfiguration.h
@@ -35,7 +35,7 @@ namespace catapult { namespace config {
 		uint8_t MaxCosignersPerAccount;
 
 		/// Maximum number of accounts a single account can cosign.
-		uint8_t MaxCosignedAccountsPerAccount;
+		uint32_t MaxCosignedAccountsPerAccount;
 
 	private:
 		MultisigConfiguration() = default;
